Make water the source loader
This is a rewrite of water so far, that makes it all a class and changes
the way you use it with gulp. It's basicly a replacement of gulp.src, as
it's making a lot more sense to start out with it, you can still merge
streams and do everything the gulp way.